+++ src/test/scala/esmeta/lang/LangTest.scala
@@ -359,54 +359,4 @@ object LangTest {
 
   // algorithm types
   lazy val ty = Type(RecordT("Base"))
-
-  def collectTypes(h: spec.Head): List[Type] = {
-    val types = MSet.empty[Type]
-    val _ = {
-      h match
-        case spec.AbstractOperationHead(isHostDefined, name, params, retTy) =>
-          for {
-            param <- params
-          } types += param.ty
-          types += retTy
-        case spec.NumericMethodHead(baseTy, name, params, retTy) =>
-          for {
-            param <- params
-          } types += param.ty
-          types += baseTy
-          types += retTy
-        case spec.SyntaxDirectedOperationHead(
-              target,
-              methodName,
-              isStatic,
-              withParams,
-              retTy,
-            ) =>
-          for {
-            param <- withParams
-          } types += param.ty
-          for {
-            param <- target.toList
-            rhsParam <- param.rhsParams
-          } types += rhsParam.ty
-          types += retTy
-        case spec.ConcreteMethodHead(concMethodName, receiver, params, retTy) =>
-          for {
-            param <- params
-          } types += param.ty
-          types += retTy
-        case spec.InternalMethodHead(methodName, receiver, params, retTy) =>
-          for {
-            param <- params
-          } types += param.ty
-          types += receiver.ty
-          types += retTy
-        case spec.BuiltinHead(path, params, retTy) =>
-          for {
-            param <- params
-          } types += param.ty
-          types += retTy
-    };
-    types.toList
-  }
 }
