--- tests/result/lang/JsonSmallTest
@@ -1,1 +1,0 @@
-lang.JsonSmallTest: 325 / 325
