--- tests/result/lang/StringifySmallTest
@@ -1,1 +1,0 @@
-lang.StringifySmallTest: 574 / 574
