+++ tests/result/lang/TypeSmallTest
@@ -1,0 +1,1 @@
+lang.TypeSmallTest: 1 / 1
