+++ tests/result/lang/TypeSmallTest.json
@@ -1,0 +1,3 @@
+{
+  "Stringify" : true
+}
